almacenamos la profundidad maxima en GA_P para poder leerla como informacion en cualquier sitio + cambiamos los valores de la clase nodo, ahora son ints, que se referirán a posiciones en el cromosoma

+++ include/nodo.hpp
@@ -46,7 +46,7 @@ struct Nodo {
 	  * @brief Valor asociado al nodo
 	  */
 
-	double valor;
+	int valor;
 
 	/**
 	  * @brief Establecer el tipo del nodo a un operador aleatorio entre los posibles operadores.
